field_provider.py mutable default argument (#187)
Hello,

in the FieldProvider class, there are multiple instances of mutable
default arguments.
I have observed, that when using the FieldFriend and creating a field,
adding more than one row results in the rows sharing the same points
until restarted.
The root cause are mutable default arguments (the list of points given
when calling create_row as example), which are then referenced in 2
different instances of Row. When restarting the the program, the rows
are restored from a file, in which case, the list of points are
instantiated seperately (fixing this shared reference).

The fix is to simply change the function to explicitly create a new
list, when no argument is supplied.