Field Creation Wizard and Field Class Changes (#183)
In the course of planning the new field navigation #180 , the wizard for
creating fields will be updated and will be accessible from the start
page (the field_planner will be deactivated temporary). In addition,
fields will be saved differently from now on. Only the AB line and a few
field parameters will be saved persistently. Rows and the perimeter of
the field will from now on be calculated dynamically from the
parameters. For now we concentrate on only one field for one robot.
Also, obstacles will be added in future and are neglected.

Todos:

- [x]  update Field Creation Wizard and move it to the main page
- [x] update Field class (save needed params and calculate rows and
boundary)
- [x] deactivate field_planner
- [x] visualize the field in leaflet
- [x] fix rows and boundary calculation issue
- [x] update tests

---------

Co-authored-by: Pascal Schade <165774906+pascalzauberzeug@users.noreply.github.com>
Co-authored-by: Pascal Schade <pascal@zauberzeug.com>