Support Points for Specific Rows (#200)
Currently the rows of a field are calculated with the first row (the
AB-line), the row-spacing and the number of rows. This PR adds the
ability to set support points for a single line that does not fit the
given row-spacing.

Example:
The created field has a AB-line a row spacing of 50cm and 10 rows.
Though, the spacing between the fourth and fifth row is 70cm. By placing
the robot on the fifth row and adding a support point for this row the
calculation of all 10 rows now includes the support points. The
following rows are based on the position of the fifth row but with the
given row-spacing of 50cm.

- [x] adding support point dialog
- [x] store support points in field
- [x] calcultate rows with support points
- [x] calculate boundary based on these novel rows
- [x] reset a support point for a specific row that already has a support point
- [x] write tests
- [x] test on a real system

---------

Co-authored-by: Pascal Schade <165774906+pascalzauberzeug@users.noreply.github.com>
Co-authored-by: Pascal Schade <pascal@zauberzeug.com>