Field with Beds (#219)
After creating a simple field with row number, row spacing and buffer
width this PR now adds the possibility to specify beds. By simple
indicating the number of beds and the spacing between those, a field can
how have beds. The row number will then be the number of rows per bed.
Safety points still work as before and the calculation of the further
rows are based on those.

Todo:

- [x] add beds in field creator
- [x] add beds in Field class (including persistence)
- [x] edit the row calculation
- [x] make beds editable
- [x] write tests
- [x] test on a robot

---------

Co-authored-by: Pascal Schade <pascal@zauberzeug.com>