No Bumper Handling (#241)
When the robot has no bumper the automation_watch will throw an
assertion as you can see in Issue #240. There are Fieldfriends that do
not have bumpers at all. That is why we need an appropriate handling of
it in the automation_watch. We would like to log a warning in case a
robot with bumpers has an error or disconnected bumpers but beside that
we just deactivate the bumper watcher.