Handling Restoring Fields with Errors (#243)
Currently fields that have incomplete or faulty data which may happen to
updates in the code base, the fields are getting deleted and will not be
stored again. We want to prevent this. When data is loaded from the
persistence missing values should be replaced to a default value and the
user should be getting informed by a notification to recheck the field's
data.