Add EntityLocator base class (#287)
This pull request introduces an EntityLocator base class that serves as
a foundation for future classes that use sensor data to locate objects
in 3d space. The existing PlantLocator has been refactored to inherit
from this new base class.

And I moved the ui from the operations block to the dev page