Fetch detector info if not available (#294)
As described in https://github.com/zauberzeug/field_friend/issues/284,
the crop categories are only fetched once on startup. If the detector is
not available in this moment, the categories will be empty and no crops
can be detected.

This PR fetches them on startup and at the start of an automation.
Additionaly I replaced the self built request by
[detector.fetch_model_version_info](https://github.com/zauberzeug/rosys/blob/2c08346da93eccf908657358be274268a4c512a1/rosys/vision/detector.py#L144)
and the current and target model versions are displayed in the dev ui.

@LukasBaecker Like this we could only select a cultivated crop if we
start and stop the automation once if the detector was not available
before. We could check regularly if detector_info is None. What do you
think

---------

Co-authored-by: Lukas Baecker <baecker.lukas@outlook.de>