Use tool offset in can_reach (#280)
All tools are at least slightly off the center of the robot and that has
to be taken in account when weeding.
Before, the work offsets were added in the puncher, but not in the
`can_reach` function, which led to errors when the robot drove to a weed
that it in the end couldn't reach.