Add low bandwidth page (#313)
When checking our cameras, we noticed that the main page uses up to 250
MB of data per hour. To reduce this, we removed the camera, the Leaflet
map, and the robot scene. To still have some idea of what the robot is
doing, we are adding the log view and state tracking to this page

---------

Co-authored-by: Pascal Schade <pascal@zauberzeug.com>