Fix on spot turning accuracy (#332)
### Motivation

We noticed that the robot had a small offset from the row, when it
started weeding on a new row. We first thought, that the drive to the
row was the problem, that's why we started
https://github.com/zauberzeug/field_friend/pull/283.
And while there is a problem with stopping correctly, this issue is
because the target linear velocity is set to 0.5.

### Implementation

It was fixed by setting the linear velocity to 0. I added a test for it
before fixing it, to proof it.
It was already tested as part of
https://github.com/zauberzeug/field_friend/pull/283

### Progress

- [x] I chose a meaningful title that completes the sentence: "If
applied, this PR will..."
- [x] I chose meaningful labels (if GitHub allows me to so).
- [x] The implementation is complete.
- [x] Tests with a real hardware have been successful (or are not
necessary).
- [x] Pytests have been added (or are not necessary).
- [x] Documentation has been added (or is not necessary).