Make the Canopen axis more robust (#331)
**Motivation**

When doing testing on the Canopen axis of `u5` we noticed, that the
`z_axis` will sometimes not move upwards after it has drilled. When we
did some further investigation we also saw, that sometimes the
referncing will not work, when the robot has just started up. This
happend without any feedback, as to why it did not work.

**Implementation**

To combat these errors we made the axis more robust with the usage of
asserts and handling the static wait times in while loops. The while
loops were chosen, in order to make shure that all the nececarry steps
are taken in the right order, with dynamic wait times. We are now
checking, if the `robot_brain.is_ready`, the motor is `initalized` and
if the motor `is_operational`. These checks are made to ensure, that the
hardware is ready to work.

**Progress**

- [x] implementation is complete
- [x] tested on u5
- [x] tested in simulation