Disable `enable_esp_on_startup` by default (#326)
While testing [F21](https://github.com/zauberzeug/field_friend/pull/315)
we noticed that the Robot Brain did enable the ESP after the start of
RoSys, which should not be the case for a new Robot Brain, since they
enable it automatically.


[Currently](https://github.com/zauberzeug/rosys/blob/e3524d9c23d721fa797b330c55d02739bb4f2ce7/rosys/hardware/robot_brain.py#L26)
the default is `True`, but it will be `False` for all upcoming Robot
Brains, so we can set it to that.

@Johannes-Thiel can you check which robots don't enable it
automatically? I think F15 is the first one that does.
And please test this with PR with an up to date Robot Brain.

I updated U4, U5 and U6. I think F12 and F13 should get
`enable_esp_on_startup=True` aswell.