Improve navigating to target and Tornado implement handling (#283)
While investigating issues with U4's tornado implement, we discovered
that the previous navigation wasn't handling the robot's deceleration
properly, which lead to a rework of the whole stack. Alongside with
several minor changes to the weeding process.

ToDos:
- [x] Block implement in specific navigation stages
- [x] Handle Work-Offset parameters in navigation (that's why the
weeding tests for U6 fail)
- [x] ~Use specific work offsets for every implement and add it to the
calculation in `get_move_target`~ -> Later with implement reworks
- [x] Add fixture and test for `WheelsSimulationWithAcceleration`
- [x] Handle acceleration for close targets
- [x] Wait for PRs:
  - https://github.com/zauberzeug/rosys/pull/298
  - https://github.com/zauberzeug/rosys/pull/297
- [x] Real life tests

---------

Co-authored-by: Lukas Baecker <baecker.lukas@outlook.de>
Co-authored-by: Johannes-Thiel <jojo.thadfun@gmail.com>