Manually use garbage collection (#336)
### Motivation
When using the normal rosys garbage collection all functions will be
stopped each minute. To prevent this we are calling the garbage
collection manually, before enabling the implement when starting a new
row and after the automation is finished. If no atomation is running we
are starting the garbage collection every 5 minutes.

### Implementation
We choose to use the garbage collection in the `field_navigation.py` at
this point, because the robot is standing still while the implement will
be enabled. For this reason a second of additional waiting time is okay.
### Progress

- [x] I chose a meaningful title that completes the sentence: "If
applied, this PR will..."
- [x] I chose meaningful labels (if GitHub allows me to so).
- [x] The implementation is complete.
- [x] Tests with a real hardware have been successful (or are not
necessary).
- [x] Pytests have been added (or are not necessary).
- [x] Documentation has been added (or is not necessary).
- [x] Wait for https://github.com/zauberzeug/rosys/pull/300
- [x] Also wait for https://github.com/zauberzeug/rosys/pull/302

edit: the commit message "update to RoSys 0.25.1" is not correct, it is
an update to 0.25.0

---------

Co-authored-by: Pascal Schade <pascal@zauberzeug.com>