Reduce soil damage with a three-point row change (#351)
### Motivation

We noticed that turning the robot on the spot when changing rows is
quite destructive to wet or loose soil. It also carries quite a lot of
soil into the tracks of the field friend. @SeaTechRC thought of a
three-point-turn to change rows, thanks for that!

### Implementation

Like the space efficient way to turn a car, the robot will now do a
three step manouver. The wanted radius can be set in the field
navigation options.

### Progress

- [x] I chose a meaningful title that completes the sentence: "If
applied, this PR will..."
- [x] I chose meaningful labels (if GitHub allows me to so).
- [x] The implementation is complete.
    - [x] Add three-point-turn
    - [x] Fix formatting
    - [x] Make some logging more clear
    - [x] Add ruff to the developer requirements
- [x] For some reason, the tests fail with a radius of 1.5m, but work
perfectly fine with 1.499 and 1.501
    - [x] Don't allow a bigger radius than your field buffer
    - [x] Remove old turn and the developer option
- [x] Tests with a real hardware have been successful (or are not
necessary).
- [x] Pytests have been added (or are not necessary).
- [x] Documentation has been added (or is not necessary).

---------

Co-authored-by: Johannes-Thiel <jojo.thadfun@gmail.com>