Introduce buttons to capture images from the cameras for upload into the Learning Loop (#359)
### Motivation

In new scenarios we need to add a first set of training data. Currently
there is no way to trigger a capturing and upload from specific cameras.

### Implementation

I extracted the capture code from #315 (added by @pascalzauberzeug) and
made it reusable by enclosing it in a separate class. The capture
actions can be triggered on the Circle Sight page and via two extra
buttons in the Zauberzeug App.

### Progress

- [x] I chose a meaningful title that completes the sentence: "If
applied, this PR will..."
- [x] I chose meaningful labels (if GitHub allows me to so).
- [x] The implementation is complete.
- [x] Wait for https://github.com/zauberzeug/field_friend/pull/315, that
will also fix the pylint error
- [x] Tests with a real hardware have been successful.
- [x] Pytests are not necessary.
- [x] Documentation is not necessary.

---------

Co-authored-by: Pascal Schade <pascal@zauberzeug.com>