Remove the BatteryWatcher class (#365)
### Motivation & Implementation

The `BatteryWatcher`'s only job was to monitor the battery and call the
`release_battery_relay` function of the `BatteryControlHardware` once on
startup and when the battery stops charging.

I deleted the class and moved the jobs to `FieldFriendHardware` where
`BatteryControlHardware` is created.
For this I added the `CHARGING_STARTED` and `CHARGING_STOPPED` events to
the BMS hardware module.

### Progress

- [x] I chose a meaningful title that completes the sentence: "If
applied, this PR will..."
- [x] I chose meaningful labels (if GitHub allows me to so).
- [x] The implementation is complete.
     - [x] Wait for https://github.com/zauberzeug/rosys/pull/323
- [x] Tests with a real hardware have been successful (or are not
necessary).
- [x] Pytests have been added (or are not necessary).
- [x] Documentation has been added (or is not necessary).