Use RobotLocator's position on LeafletMap (#386)
### Motivation & Implementation

The LeafletMap was using the GNSS measurement position directly, which
caused a jumpy icon position but more importantly made it hard to detect
problems with the Kalman Filter or the GNSS reference.

I also reduced the map update interval to a 5 second timer. Because the
GNSS event comes in every 0.1 seconds, which is not necessary.

### Progress

- [x] I chose a meaningful title that completes the sentence: "If
applied, this PR will..."
- [x] I chose meaningful labels (if GitHub allows me to so).
- [x] The implementation is complete.
- [x] Tests with a real hardware have been successful (or are not
necessary).
- [x] Pytests have been added (or are not necessary).
- [x] Documentation has been added (or is not necessary).