Fix robot between rows test (#387)
### Motivation & Implementation

This PR fixes https://github.com/zauberzeug/field_friend/issues/381.
Contrary to my first idea, the test was failing due to simulated GNSS
noise, because it will sometimes move the robot just slightly over the
threshold. This is now fixed (we still need the noise to keep the Kalman
Filter stable in longer tests though).

The mentioned not awaited coroutines were not the issue, but just an
additional error that is fixed with
https://github.com/zauberzeug/rosys/pull/334.

### Progress

- [x] I chose a meaningful title that completes the sentence: "If
applied, this PR will..."
- [x] I chose meaningful labels (if GitHub allows me to so).
- [x] The implementation is complete.
    - [x] Wait for https://github.com/zauberzeug/rosys/pull/334
- [x] Tests with a real hardware have been successful (or are not
necessary).
- [x] Pytests have been added (or are not necessary).
- [x] Documentation has been added (or is not necessary).