Make FieldNavigation's turn radius configurable (#390)
### Motivation & Implementation

On loose, sandy soil the turn radius of 1.5 meters seems too small. The
tracks of the robot will dig into the ground, especially with the old
track model.

While I was at it, I moved `generate_three_point_turn` to file of
`WaypointNavigation`, because it is also interesting for other field
friend based projects.


### Progress

- [x] I chose a meaningful title that completes the sentence: "If
applied, this PR will..."
- [x] I chose meaningful labels (if GitHub allows me to so).
- [x] The implementation is complete.
- [x] Tests with a real hardware have been successful (or are not
necessary).
- [x] Pytests have been added (or are not necessary).
- [x] Documentation has been added (or is not necessary).