Fix wheel_distance for all robots (#398)
### Motivation & Implementation

We noticed that the configured width of our robots was not updated
correctly.

So I measured and edited the correct values for all configs, fixed the
3D scene object and moved the `m_per_tick` to the configuration.

Both the `MeasurementsConfiguration` and `FieldFriendHardware` have to
be reworked and cleaned up some day, but that is not important for now.

### Progress

- [x] I chose a meaningful title that completes the sentence: "If
applied, this PR will..."
- [x] I chose meaningful labels (if GitHub allows me to so).
- [x] The implementation is complete. 
- [x] Tests with a real hardware have been successful (or are not
necessary).
- [x] Pytests have been added (or are not necessary).
- [x] Documentation has been added (or is not necessary).