Rework detector outbox mode handling (#400)
### Motivation & Implementation

Before we were setting the outbox mode of the detector container on a
fixed interval, which made it hard to read the logs of the detector.
I changed it so it's only setting the outbox mode on change and on
startup.
I also removed the old settings.py that was only used for allowing the
upload of images via a mobile network. Now there is a checkbox in the
plant locator ui.

To fix a mypy complaint, I also reworked `add_chip` in the
`PlantLocator` ui.

### Progress

- [x] I chose a meaningful title that completes the sentence: "If
applied, this PR will..."
- [x] I chose meaningful labels (if GitHub allows me to so).
- [x] The implementation is complete.
- [x] Tests with a real hardware have been successful (or are not
necessary).
- [x] Pytests have been added (or are not necessary).
- [x] Documentation has been added (or is not necessary).